add restiction for sigle user, add slug for paramters, add relation user and travel

+++ app/Http/Controllers/Admin/TravelController.php
@@ -8,6 +8,8 @@ use App\Http\Requests\StoreTravelRequest;
 use App\Http\Requests\UpdateTravelRequest;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TravelController extends Controller
 {
@@ -16,17 +18,20 @@ class TravelController extends Controller
      */
     public function index()
     {
-        // variabile con tutti i viaggi nel db
-        $travels = Travel::orderByDesc('id')->paginate(5);
+        // variabile che salva l'id dell'utente che si e loggato
+        $id = Auth::id();
+
+        // variabile con tutti i viaggi nel db in ordine discendente in base chi si e loggato
+        $travels = Travel::where('user_id', $id)->orderByDesc('id')->paginate(5);
 
         // rispedisce alla pagina inde di travel con tutti i viaggi
-        return view('admin.travels.index', compact('travels'));
+        return view('admin.travels.index', compact('travels', 'id'));
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create($id)
     {
         // rispedisce alla pagina creazione di travel
         return view('admin.travels.create');
@@ -64,6 +69,9 @@ class TravelController extends Controller
         //salva lo slug nella nella key slug 
         $val_data['slug'] = $slug;
 
+        // salvo l'user_id
+        $val_data['user_id'] = Auth::id();
+
         // crea un nuovo viaggio e lo inserisce nel db
         $travel = Travel::create($val_data);
 
@@ -76,8 +84,12 @@ class TravelController extends Controller
      */
     public function show(Travel $travel)
     {
-        // rispedisce alla pagina singola di un travel
-        return view('admin.travels.show', compact('travel'));
+        // se l'id dell'utente e uguale a quello del viaggio
+        if (Gate::allows('travel_checker', $travel)) {
+            // rispedisce alla pagina singola di un travel
+            return view('admin.travels.show', compact('travel'));
+        } //in caso ti esce errore 
+        abort(403, "Non hai l'autorizzazione per accedere a questa pagina");
     }
 
     /**
@@ -85,8 +97,12 @@ class TravelController extends Controller
      */
     public function edit(Travel $travel)
     {
-        // rispedisce alla pagina di modifica
-        return view('admin.travels.edit', compact('travel'));
+        // se l'id dell'utente e uguale a quello del viaggio
+        if (Gate::allows('travel_checker', $travel)) {
+            // rispedisce alla pagina di modifica
+            return view('admin.travels.edit', compact('travel'));
+        } //in caso ti esce errore 
+        abort(403, "Non hai l'autorizzazione per accedere a questa pagina");
     }
 
     /**
@@ -94,44 +110,48 @@ class TravelController extends Controller
      */
     public function update(UpdateTravelRequest $request, Travel $travel)
     {
-        // variabile con i dati validati
-        $val_data = $request->validated();
+        // se l'id dell'utente e uguale a quello del viaggio
+        if (Gate::allows('travel_checker', $travel)) {
+            // variabile con i dati validati
+            $val_data = $request->validated();
 
-        // se ce image
-        if ($request->has('image')) {
+            // se ce image
+            if ($request->has('image')) {
 
-            // se esiste l'immagine di travel
-            if ($travel->image) {
+                // se esiste l'immagine di travel
+                if ($travel->image) {
 
-                // l'immagine viene cancellata da storage
-                Storage::disk('public')->delete($travel->image);
+                    // l'immagine viene cancellata da storage
+                    Storage::disk('public')->delete($travel->image);
+                }
+                // imagine viene inserita in storage
+                $val_data['image'] = Storage::disk('public')->put('uploads/images', $val_data['image']);
             }
-            // imagine viene inserita in storage
-            $val_data['image'] = Storage::disk('public')->put('uploads/images', $val_data['image']);
-        }
 
-        // variabile che fa checker o salva il numero di travel con lo stesso nome
-        $slug_checker = Travel::where('name', $val_data['name'])->count();
+            // variabile che fa checker o salva il numero di travel con lo stesso nome
+            $slug_checker = Travel::where('name', $val_data['name'])->count();
 
-        // se esiste lo slug
-        if ($slug_checker) {
+            // se esiste lo slug
+            if ($slug_checker) {
 
-            // variabile che salva lo slug
-            $slug = Str::slug($val_data['name'], '-') . '-' . $slug_checker + 1;
-        } else {
+                // variabile che salva lo slug
+                $slug = Str::slug($val_data['name'], '-') . '-' . $slug_checker + 1;
+            } else {
 
-            // variabile che salva lo slug
-            $slug = Str::slug($val_data['name'], '-');
-        }
+                // variabile che salva lo slug
+                $slug = Str::slug($val_data['name'], '-');
+            }
 
-        //salva lo slug nella nella key slug 
-        $val_data['slug'] = $slug;
+            //salva lo slug nella nella key slug 
+            $val_data['slug'] = $slug;
 
-        // modifica travel con i nuovi dati
-        $travel->update($val_data);
+            // modifica travel con i nuovi dati
+            $travel->update($val_data);
 
-        // rispedisce alla pagina index di travels
-        return to_route('admin.travels.index')->with('message', "Hai modificato il viaggio: $travel->name");
+            // rispedisce alla pagina index di travels
+            return to_route('admin.travels.index')->with('message', "Hai modificato il viaggio: $travel->name");
+        } //in caso ti esce errore 
+        abort(403, "Non hai l'autorizzazione per accedere a questa pagina");
     }
 
     /**
@@ -139,11 +159,25 @@ class TravelController extends Controller
      */
     public function destroy(Travel $travel)
     {
-        if ($travel->image) {
-            Storage::disk('public')->delete($travel->image);
-        }
-        $name = $travel->name;
-        $travel->delete();
-        return redirect()->back()->with('message', "Hai cancellato il viaggio: $name");
+        // se l'id dell'utente e uguale a quello del viaggio
+        if (Gate::allows('travel_checker', $travel)) {
+
+            // se esiste l'immagine di travel
+            if ($travel->image) {
+
+                // l'immagine viene cancellata da storage
+                Storage::disk('public')->delete($travel->image);
+            }
+
+            // variabile che salva il nome di travel+
+            $name = $travel->name;
+
+            // cancello travel
+            $travel->delete();
+
+            // ritorna alla pagina index
+            return redirect()->back()->with('message', "Hai cancellato il viaggio: $name");
+        } //in caso ti esce errore 
+        abort(403, "Non hai l'autorizzazione per accedere a questa pagina");
     }
 }
